refactor: split feishu helpers and tests
