+++ extensions/feishu/src/monitor.account.ts
@@ -20,6 +20,7 @@ import {
   warmupDedupFromDisk,
 } from "./dedup.js";
 import { isMentionForwardRequest } from "./mention.js";
+import { applyBotIdentityState, startBotIdentityRecovery } from "./monitor.bot-identity.js";
 import { fetchBotIdentityForMonitor } from "./monitor.startup.js";
 import { botNames, botOpenIds } from "./monitor.state.js";
 import { monitorWebhook, monitorWebSocket } from "./monitor.transport.js";
@@ -619,70 +620,6 @@ function registerEventHandlers(
   });
 }
 
-// Delays must be >= PROBE_ERROR_TTL_MS (60s) so each retry makes a real network request
-// instead of silently hitting the probe error cache.
-const BOT_IDENTITY_RETRY_DELAYS_MS = [60_000, 120_000, 300_000, 600_000, 900_000];
-
-export function waitForAbortableDelay(
-  delayMs: number,
-  abortSignal?: AbortSignal,
-): Promise<boolean> {
-  if (abortSignal?.aborted) {
-    return Promise.resolve(false);
-  }
-
-  return new Promise((resolve) => {
-    const timer = setTimeout(() => {
-      abortSignal?.removeEventListener("abort", handleAbort);
-      resolve(true);
-    }, delayMs);
-    timer.unref?.();
-
-    const handleAbort = () => {
-      clearTimeout(timer);
-      resolve(false);
-    };
-
-    abortSignal?.addEventListener("abort", handleAbort, { once: true });
-  });
-}
-
-async function retryBotIdentityProbe(
-  account: ResolvedFeishuAccount,
-  accountId: string,
-  runtime: RuntimeEnv | undefined,
-  abortSignal: AbortSignal | undefined,
-): Promise<void> {
-  const log = runtime?.log ?? console.log;
-  const error = runtime?.error ?? console.error;
-  for (let i = 0; i < BOT_IDENTITY_RETRY_DELAYS_MS.length; i++) {
-    if (abortSignal?.aborted) return;
-    const delayElapsed = await waitForAbortableDelay(BOT_IDENTITY_RETRY_DELAYS_MS[i], abortSignal);
-    if (!delayElapsed) {
-      return;
-    }
-    const identity = await fetchBotIdentityForMonitor(account, { runtime, abortSignal });
-    if (identity.botOpenId) {
-      botOpenIds.set(accountId, identity.botOpenId);
-      if (identity.botName?.trim()) {
-        botNames.set(accountId, identity.botName.trim());
-      }
-      log(
-        `feishu[${accountId}]: bot open_id recovered via background retry: ${identity.botOpenId}`,
-      );
-      return;
-    }
-    const nextDelay = BOT_IDENTITY_RETRY_DELAYS_MS[i + 1];
-    error(
-      `feishu[${accountId}]: bot identity background retry ${i + 1}/${BOT_IDENTITY_RETRY_DELAYS_MS.length} failed` +
-        (nextDelay ? `; next attempt in ${nextDelay / 1000}s` : ""),
-    );
-  }
-  error(
-    `feishu[${accountId}]: bot identity background retry exhausted; requireMention group messages may be skipped until restart`,
-  );
-}
-
 export type BotOpenIdSource =
   | { kind: "prefetched"; botOpenId?: string; botName?: string }
   | { kind: "fetch" };
@@ -705,26 +642,11 @@ export async function monitorSingleAccount(params: MonitorSingleAccountParams): 
     botOpenIdSource.kind === "prefetched"
       ? { botOpenId: botOpenIdSource.botOpenId, botName: botOpenIdSource.botName }
       : await fetchBotIdentityForMonitor(account, { runtime, abortSignal });
-  const botOpenId = botIdentity.botOpenId;
-  const botName = botIdentity.botName?.trim();
-  botOpenIds.set(accountId, botOpenId ?? "");
-  if (botName) {
-    botNames.set(accountId, botName);
-  } else {
-    botNames.delete(accountId);
-  }
+  const { botOpenId } = applyBotIdentityState(accountId, botIdentity);
   log(`feishu[${accountId}]: bot open_id resolved: ${botOpenId ?? "unknown"}`);
 
-  // When the startup probe failed, retry in the background so the degraded
-  // state (responding to all group messages) is bounded rather than permanent.
   if (!botOpenId && !abortSignal?.aborted) {
-    log(
-      `feishu[${accountId}]: bot open_id unknown; starting background retry (delays: ${BOT_IDENTITY_RETRY_DELAYS_MS.map((d) => `${d / 1000}s`).join(", ")})`,
-    );
-    log(
-      `feishu[${accountId}]: requireMention group messages stay gated until bot identity recovery succeeds`,
-    );
-    void retryBotIdentityProbe(account, accountId, runtime, abortSignal);
+    startBotIdentityRecovery({ account, accountId, runtime, abortSignal });
   }
 
   const connectionMode = account.config.connectionMode ?? "websocket";
