+++ src/agents/pi-embedded-runner/model.ts
@@ -2,7 +2,6 @@ import type { Api, Model } from "@mariozechner/pi-ai";
 import type { AuthStorage, ModelRegistry } from "@mariozechner/pi-coding-agent";
 import type { OpenClawConfig } from "../../config/config.js";
 import type { ModelDefinitionConfig } from "../../config/types.js";
-import { normalizeGoogleApiBaseUrl } from "../../infra/google-api-base-url.js";
 import {
   clearProviderRuntimeHookCache,
   prepareProviderDynamicModel,
